ImportNewSongsButton and SearchBar
